+++ src/Hpack/License.hs
@@ -1,21 +1,17 @@
 {-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE ViewPatterns #-}
+{-# LANGUAGE LambdaCase #-}
 module Hpack.License where
 
 import           Control.Arrow ((&&&))
-import           Data.List
-import           Data.Ord (comparing)
-import           Data.Text (Text)
-import qualified Data.Text as T
-import           Data.Text.Metrics
 
 import           Distribution.Pretty (prettyShow)
+import           Distribution.Version (mkVersion)
 import qualified Distribution.License as Cabal
 import qualified Distribution.SPDX.License as SPDX
-import           Distribution.SPDX.LicenseId
 import           Distribution.Parsec.Class (eitherParsec)
 
-import           Hpack.SpdxLicenses (licenses)
+import qualified Data.License.Infer as Infer
 
 data License a = DontTouch String | CanSPDX Cabal.License a | MustSPDX a
   deriving (Eq, Show, Functor)
@@ -33,21 +29,16 @@ parseLicense license = case lookup license knownLicenses of
     spdxLicense :: Maybe SPDX.License
     spdxLicense  = either (const Nothing) Just (eitherParsec license)
 
-probabilities :: Text -> [(LicenseId, Double)]
-probabilities license = map (fmap probability) licenses
+inferLicense :: String -> Maybe (License SPDX.License)
+inferLicense = fmap (uncurry CanSPDX . (id &&& Cabal.licenseToSPDX) . toLicense) . Infer.inferLicense
   where
-    probability = realToFrac . levenshteinNorm license
-
-inferLicense :: String -> Maybe (License String)
-inferLicense (T.pack -> xs) = case maximumBy (comparing snd) (probabilities xs) of
-  (license, n) | n > 0.85 -> Just (toLicense license)
-  _ -> Nothing
-  where
-    toLicense :: LicenseId -> License String
-    toLicense license = (case license of
-      MIT -> CanSPDX Cabal.MIT
-      BSD_2_Clause -> CanSPDX Cabal.BSD2
-      BSD_3_Clause -> CanSPDX Cabal.BSD3
-      BSD_4_Clause -> CanSPDX Cabal.BSD4
-      _ -> MustSPDX
-      ) (licenseId license)
+    toLicense = \ case
+      Infer.MIT -> Cabal.MIT
+      Infer.BSD2 -> Cabal.BSD2
+      Infer.BSD3 -> Cabal.BSD3
+      Infer.BSD4 -> Cabal.BSD4
+      Infer.GPLv2 -> Cabal.GPL (Just $ mkVersion [2])
+      Infer.GPLv3 -> Cabal.GPL (Just $ mkVersion [3])
+      Infer.LGPLv2_1 -> Cabal.LGPL (Just $ mkVersion [2,1])
+      Infer.LGPLv3 -> Cabal.LGPL (Just $ mkVersion [3])
+      Infer.AGPLv3 -> Cabal.AGPL (Just $ mkVersion [3])