+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## Next
+## Changes in 0.29.3
   - Desugar `^>=` for compatibility with `Cabal < 2` (see #213)
+  - Add support for GPL, LGPL and AGPL when inferring `license` (see #305)
 
 ## Changes in 0.29.2
   - Add missing `extra-source-files` (see #302)